Upgrade Typo to run with Rails 0.14.1 (AKA 1.0rc2)
This required a few minor changes, and I took the opportunity to do a bit of
refactoring that I'd been avoiding for a while--text filters no longer use
render() to do their filtering.  That was broken with 1.0, and I've been
meaning to fix it anyway.

All tests now pass for me, but I haven't tested it extensively outside of
the unit test system, so be careful.

This may or may not work with 0.13.1--there's no reason why it shouldn't, but
it's untested.  I'm going to include a few 1.0rc2-specific cleanups soon, so
the days of the trunk working with 0.13.1 are numbered.

git-svn-id: http://svn.typosphere.org/typo/trunk@685 820eb932-12ee-0310-9ca8-eeb645f39767