Rails 1.1 rocks my world!
'/admin/sidebar' just got some RJS magic sprinkled on it so you get a 
spinner in the 'active sidebars' header while the backend is updating
stuff. Coming soon: Disabling the 'publish changes' button during loading.

However, it is late, I am tired and I must be up at <ungodly>.30 tomorrow
so I leave it as an exercise for the interested hacker.

git-svn-id: http://svn.typosphere.org/typo/trunk@997 820eb932-12ee-0310-9ca8-eeb645f39767