I continue to pick away at the Sidebar system. In this patch we:
  * Pull Sidebars::Plugin out into lib
  * Add a Sidebars::!ComponentPlugin which all component based sidebars
    inherit from (not that there are any other sorts of sidebar yet. Bear 
    with me
  * Pulled !SidebarController out of components and into app/controllers
  * Lay groundwork for eliminating {{{staged_position}}} from the Sidebar 
    model

We've also got the beginnings of a sidebar preview here, /sidebar will 
show you a dummy article with a sidebar -- eventually this should work
to show you a preview of the staged sidebar configuration.

Coming up, a way of calling actions other than 'configure' and 'content'
on your sidebars -- a necessity for advanced navigation tools for 
instance.

git-svn-id: http://svn.typosphere.org/typo/trunk@1002 820eb932-12ee-0310-9ca8-eeb645f39767