I've havered over doing this for a while now, if only because Rails is very strongly of the opinion that models don't need to know about url_for.
Which is all very well, but Blog is a model of something that has
contents accessible through urls. 

As always in these matters, the true test of whether something's a
good idea or not is:

  Does it make the code clearer?

And in this case, I think it does. By adding Blog#url_for, and 
especially by allowing it to take an object that responds to 
{{{location(anchor=nil, only_path=true)}}} I've managed to 
refactor out a great deal of repetition.

Since url_for is also the stumbling block which stops us turning 
textfilters back into models, you can also consider this as a 
step along the road to making textfilters a little less weird.

git-svn-id: http://svn.typosphere.org/typo/trunk@1036 820eb932-12ee-0310-9ca8-eeb645f39767