Web notifications (trackbacks, pingbacks that stuff) are now triggered from the WebNotifier content observer.
Note that we've now added a new setting to Blog 'default_server_url' as a 
fallback for finding the blog's base url. If you're experiencing problems
(I don't think you should be, it's only currently an issue in the tests and
our blog fixtures has an appropriate url set), go to your /admin page and 
set the default url there (it should pick up a sensible default, so you can
just check that for sanity and hit save if you like).

git-svn-id: http://svn.typosphere.org/typo/trunk@1046 820eb932-12ee-0310-9ca8-eeb645f39767