An installer for Typo.
This allows two-step installs from a Typo gem.  I'll upload a test GEM shortly.

I had to remove the test/mocks/themes/azure symlink and replace it with a partial copy
because it did bad things to the GEM, and that broke tests.  I'd love to see a better
workaround.

git-svn-id: http://svn.typosphere.org/typo/trunk@1068 820eb932-12ee-0310-9ca8-eeb645f39767