Fix a major memory leak.  It went like this:
before_filter A (allocate something)
 render page
after_filter A (free it)

Now, let's add a new filter, filter B

before_filter A (allocate something)
 before_filter B
   render page
 after_filter B
after_filter A (free it)

Filter B is the action cache.  The after filter throws 
pages into the cache, and the before filter checks
for cache hits.  If the requested page is found, then
it'll return it and cancel the rest of the rendering
process.

Which means that after_filter A will never run.

Which leaks RAM like a sieve.

The fix, for now, is to use prepend_around_filter instead
of the default append_around_filter.  Longer-term, I'm not
sure what's the best fix, because we just make
authentication tricky.

Closes #972.

git-svn-id: http://svn.typosphere.org/typo/trunk@1088 820eb932-12ee-0310-9ca8-eeb645f39767