Big changes to the state engine for feedback. Feedback now starts off in the 'Unclassified' state and moves to either JustPresumedHam or PresumedSpam depending on the results of Feedback#classify (the various Just* states are there as one shot states to handle sending of notifications or registering of spam/ham classification with akismet). JustPresumedHam decays to PresumedHam as soon as the feedback has been saved and the notifications sent.
And there they rest until the classifications are solidified by the blog
administrator on the feedback page. Feedback ends up as either Spam or Ham
with an intermediate 'JustMarkedAs*' step for registering with Akismet.

It should be possible to do some useful filtering on the feedback page 
(for instance, only see those messages whose classification hasn't been
confirmed), but for that we need to add a couple of boolean fields to the 
contents table.

git-svn-id: http://svn.typosphere.org/typo/trunk@1188 820eb932-12ee-0310-9ca8-eeb645f39767