Who said Scott was the only person who gets to do the big patches?
This is the first step towards moving all our sidebars out into 
controller free plugins in vendor/plugins. There's still a lot of work to do
but the patch was getting huge and the tests were all passing, so it's time to
push it back to the trunk for everyone to have a play.

Changes:

* Rejigged the sidebar generator to generate a new school sidebar.
* Ported all the currently distributed sidebars to the new style
* Moved a bunch of sidebar specific models to their sidebar's directory
* In passing, killed a bunch of edge rails deprecation warnings (tests 
  still fail on the edge, but at least they fail without a bunch of 
  deprecatory noise)
* Some general housekeeping. I haven't, yet, removed all trace of the old
  sidebar controllers, but we're getting there.
* The components directory is now blissfully empty!

git-svn-id: http://svn.typosphere.org/typo/trunk@1245 820eb932-12ee-0310-9ca8-eeb645f39767