First step on a long needed refactoring. We're splitting the contents table into two. We now have a contents table, which contains Pages and Articles, and a Feedback table, which contains Comments and Trackbacks. There's still a bunch of null columns in each table, and the indexing needs tweaking, but it's a start.
NB: The migration is untested on anything but my Mysql installation and if it 
goes wrong you stand to lose all your feedback. So take a backup before you 
try it. The Migration itself throws an exception by default to remind you to do the 
backup. Comment it out once you're safe (or brave/foolhardy).

git-svn-id: http://svn.typosphere.org/typo/trunk@1302 820eb932-12ee-0310-9ca8-eeb645f39767