We've added svn:externals links to rspec and rspec_on_rails and rejigged the specs we had to work with rspec 0.9.4's style.
I expect that I shall be adding any new functionality to Typo 
in a spec first style, so if you want to be sure that any
additions you make don't break things, you need to be running
the specs.

I expect that our gems will not include the whole rspec 
framework though.


git-svn-id: http://svn.typosphere.org/typo/trunk@1440 820eb932-12ee-0310-9ca8-eeb645f39767