Right ho, /articles/category now redirects to /articles/categories, which is a resource now.
Eventually, CategoriesController will take on the duties of 
Admin::CategoriesController, but not quite yet, for now it just handles
index and show.

Note that we've also introduced a new type of resource, the 'inflected_resource'
which uses /articles/categories/... for its collection methods, but 
/articles/category/whatever for its members. This means that you we can 
have a category with a permalink of 'new' and it won't clash with with the 
admin form for creating a new category.

Isn't that lovely?


git-svn-id: http://svn.typosphere.org/typo/trunk@1510 820eb932-12ee-0310-9ca8-eeb645f39767