Fix mocking of published articles for GroupingController
Once upon a time, I suspect that GroupingController called
grouping.published_articles to generate the list of published articles.
But currently, it calls:

  @articles =
    grouping.articles.paginate(:page => params[:page],
                               :conditions => { :published => true })

This means that the specs for our three major types of groupings need to
override grouping.articles instead of grouping.published_articles when
setting up their mocks.  There's some regretable code duplication among the
various spec files, so this had to be fixed in three places.

There's also a larger philosophical issue with published_articles.
Currently, it has no callers outside of the various spec/**/*.rb files, and
it is defined in a variety of inconsistent ways.  See the following bug for
further discussion of what to do here:

  http://typosphere.org/projects/typo/issues/show?id=1294