Prevent accidentally changing the blog's title
When running the unit tests, I was getting the following error:

  Test::Unit::AssertionFailedError in 'BackendController
    test_blogger_get_users_blogs'
  <"test blog trackbacks trackbacks"> expected but was
    <"test blog">.
  ./spec/controllers/backend_controller_spec.rb:37:in
    `test_blogger_get_users_blogs'

First problem: This error message has accidentally reversed the expected and
actual values.

Second problem: The following two lines in xml_controller.rb were
accidentally changing the blog title.  This occurs because '<<' actually
updates a string in place.

  @feed_title = this_blog.blog_name
  ...
  @feed_title << " trackbacks"

The fix: Make a copy of the blog title before storing it in @feed_title.