Fix MetaWeblog API to interpret time zones consistently
The goal of this patch is to fix the following unit test failure, which
started occuring at (IIRC) 7pm Eastern Standard Time this evening:

  Test::Unit::AssertionFailedError in 'BackendController
    test_meta_weblog_new_post'
  <Fri Nov 28 05:00:00 UTC 2008> expected but was
  <Sat Nov 29 05:00:00 UTC 2008>.
  ./spec/controllers/backend_controller_spec.rb:186:in
    `test_meta_weblog_new_post'

This failure does not occur earlier in the day, at least in my time zone.

To fix this problem, we need to fix the dateCreated field returned by
getPost to use a proper dateTime value instead of a string.  This may
affect users who use MetaWeblog API clients, though hopefully it will only
affect them for the better.

The problem: The existing code is attempting to use UTC for all dateTime
values, but failing, because it sometimes returns strings where dateTime
values are expected.

Why the existing code is correct to _attempt_ to use UTC everywhere:

The XML-RPC spec does not allow serialized dateTime values to contain a
time zone, and instead relies on clients and servers to agree on a
convention for interpreting dateTime values:

  http://www.xmlrpc.com/spec

  * What timezone should be assumed for the dateTime.iso8601 type? UTC?
    localtime?

    Don't assume a timezone. It should be specified by the server in its
    documentation what assumptions it makes about timezones.

As you can see from the following discussion, the MetaWeblog API inherits
this ambiguity:

  http://www.25hoursaday.com/weblog/CommentView.aspx?guid=c210531e-646c-41e8-a81e-f3efd4ba8df4

As noted in that comment thread, most blogging tools have standardized on
using UTC everywhere, since it can be unambiguously interpreted without a
timezone.

The fixes:

  1) First, in the test cases, Time.now.midnight.utc and
     Time.now.utc.midnight are _not_ the same in most time zones.  This is
     fixed by relying on the previously computed value of
     'article.published_at' instead of checking the clock a second time.
     Doing this also allows us to work around a hypothetical test case
     failure which would occur if the clock passed midnight in the middle
     of the test case.

  2) Second, ActionWebService will simply Do The Right Thing with dates,
     but only if it can identify them as such.  Previously, the function
     MetaWeblogService.article_dto_from would actually try to serialize
     dates as strings instead of XML-RPC dateTime values.  This caused a
     variety of time zone lossage, and made the test case failures very
     hard to debug.  The biggest consequence of this bug: getPost would
     return a string instead of an XML-RPC dateTime value for dateCreated.

I've added a test case to make sure that we can round-trip dates without
corrupting them.