Fix various issues in new-style extended content handling
Once upon a time, the Typo web interface allowed you to separately edit the
"body" and the "extended content" of an article.  Now, the two different
types of text are stored in a single body field, and separated by
"<!--more-->".  This change has not yet been pushed through to the
underlying database.

This patch doesn't attempt to address any of the larger design or
implementation issues, but it does fix a number of minor problems:

  - The build_extended and get_extended methods on ContentController were
    public, which actually made them callable for users on the web.  This
    kind of helper method should generally be protected or private in a
    controller.
  - The build_extended method would silently discard data if the user
    accidentally inserted two copies of '<!--more-->' in an article.
  - The build_extended method didn't remove the newlines inserted by the
    get_extended method, which would cause the article to pick up two extra
    newlines each time it was saved.

The patch moves the merging/extracting code into the article model, fixes
the data-loss bug, makes sure that articles can be edited without any
inadvertent changes, and adds a number of unit tests to make sure that
everything works.