Fixing are a few remaining bugs in the editor I did not see at first glance because I use a very permissive crappy browser to develop (ie. Firefox).
A misclosed div breaks the editor in Safari, causing a style bug and making the tag field disappear when you switch editor.

Editor should have rounded corner, using border-radiud, moz-border-radius and webkit-border-radius

Save block style is brocken (need a line break) on page editor

On pages, optionnal settings are displayed on new post. This is the contrary of what's expected