Resolve bundler issues.
By adding a group :adapters, but telling bundler not to load it, we can
keep a uniform Gemfile and Gemfile.lock, but still load a different set
of gems based on the database configuration.

Also, the gems to load for databases other than mysql have a name that
differs from the adapter names used in config/database.yml; therefore,
explicitely map adapter names to gem names.