Fixes rake tasks in production mode.
lib/bare_migration.rb was not loaded anymore, which was breaking some migrations.
Contrary to running your application in production, a Rake task does not eager load your entire code base. So only if  is false, will the application be eager-loaded in production. And  is set to true in the :environment Rake task.

Solution was given at https://rails.lighthouseapp.com/projects/8994/tickets/2506-models-are-not-loaded-in-migrations-when-configthreadsafe-is-set