Make migrations work without loading the actual models.
In the production environment, models are not loaded by default in rake
tasks. This change makes migrations work regardless, which has the added
benefit that the real models can change completely without breaking the
migrations, without having to resort to the BareMigration system. This is
particularly fortunate, since BareMigration does not support STI, which we
use with Articles and Contents.