Makes the RSS description editable from the admin.
This feature uses the same tags format as permalink definition %tag%. Available tags are:
%author%: author name
%blog_url%: blog base url
%blog_name%: blog title
%permalink_url%: permalink url

This also does a little bit of cleaning as rss_description was duplicated in both Article and Content models. I need to find a way to add tests to this, both in the model and controller parts.

[#165 state:resolved]