Turns "Show extended on RSS" into "Hide extended on RSS".
Asking user to opt in to display the whole post in the feed instead of hiding it is one of the most stupid behavior I've seen for a while and is the exact opposite of the software default's behavior.

Included migration sets the new behavior accordingly to the blog default one.