Doing some cleaning in the CSS, they really deserved it.
I dropped the call to the RTL CSS on purpose, it will come back when I understand how that really work and how I can integrate it with our bootstrap based CSS. Plus Hebrew seems not to be maintained anymore (unless someone did it and forgot to send a pull request which is fairly possible).

Which bring something quite unrelated to my mind, but should we switch to Rails YML based i18n or to the more standard gettext?