flickr: sanitize handling of API keys and secrets.
The simple API key is not enough anymore to fetch Flickr images, the
shared secret is needed as well. Since this is a secret, it can't be
stored in the GIT repository, so instead require a configuration on
the initializer itself.

Possibly, this should be configured at runtime like it's done for
Akismet.