Semantic HTML use in Article Resources, punch up message of attached
Shortens the "Currently this article has the following resources"
heading to something less passive.

Too many uses of UL and LI, removes them and instead only use list
markup on the item resources attached to a post.

Adds short message that says there are none currently attached
instead of just a blank space that goes straight into the heading
and section of available resources to include.

The translations added are just for the word, "None".