Removes the errors settings and moves the 404 error message to the template.
Errors settings were initially added to handle many error types (at least 401, 403, 404 and even 410).

It appears that having this as a setting did not bring any value to the software and no one was using it. Also, the other 4xx error codes were never implemented.

Since we're on the way to cleaning the code, we decided to remove it.