Adds frontend short statuses support. Admin management to come with next commit
Main idea: going a little further into the POSSE (publish owne site, syndicate everywhere) things, and get mor visibility by being listed as an Indie Web project.

Instead of publishing statuses directly to Twitter / Facebook / something else, we publish them on Typo. A status is a formated, 200 or 300 characters max message. Statuses have a permalink like posts, in a /st/ hierarchy. The permalink start with the status id so we can push the same message twice.
When pushed on Twitter / FB..., we shorten them to the max size of the service posts and add the short URL to the status.

Example:
Posted on Publify:
  permalink: http://myblog/st/1-hello-this-is-a-status
  short URL: http://myblog/shortlink
  message: Hello, this is a status

Posted on Twitter:
  message: Hello this is a status (myblog shortlink)

Statuses are not listed on the main timeline.

Examples in the wild: Tantek Celik (@t) way of doing. http://tantek.com/2013/183/t1/google-reader-dead-answer-integrated-indie-aggregation

Implementation on Typo:
• Statuses inherits from Content using the body field + user's text filter + permalink (this is done and covered)
• Each users with publishing rights can create statuses
• If no Twitter / Whatever account is configured, statuses are just published on Typo.
• Twitter accounts are setup on a per user basis.
• Statuses are displayed on /author/your-name pages
• Statuses timeline is /st/, displayed by th default number of items in a list.

Open questions:
• Should statuses timeline have a RSS/Atom feed?
• Should we cache the user avatar or load it from Twitter?
• Push on a per status basis?
• Is it already lunch time?