Adds a text filter to turn @mention and #hashtags into links to Twitter inside short statuses.
This is done using the html_postprocess method included in the content model.

html_postprocess is applied when views call @content.html(:field) (usually :body). In article / pages it just returns the html, but is used in the feedback model to sanitize user posted HTML.

On a more or less related thing, I've just dug into the text filter code for the first time since I took over the project. I've been willing to add Wordpress like pre and post hooks to allow user defined plugins to change the text on the fly. This was actually already implemented into Typo with the preprocess and post process text filters. Currently, the simplest way to do is to add filters params to the existing 4 text filters (none, markdown, textile and markdown + smartypants). I'll come with an elegant way to do it, already have 2-3 ideas about that.