Adds support for link auto extraction for short statuses. Works with http, https, mailto and gopher (yeah gopher is still cool).
For this, I had to add a new html_preprocess method into the content_model that's called right before the text filter so URLs with _something_ inside them won't be transformed into html before they're turned into a link.

When using markdown or textile, displayed URLs with _ something _ are still broken, but OK inside the link. Maybe the solution would have been to keep this postprocess and replace the html tags with _. I'll think about that and provide a fix.