Rename statuses as notes.
Indie Web "Standards" use the term "note" for "status" as it is less linked to a particular silo and does not have the timely meanning.

Comes with a migration to move statuses to notes. Should I add one to rebuild all the perma short links as well?

Also fixes the push to Twitter that was broken.