Add humans txt settingso
Why a TXT?
ecause it's something simple and fast to create. Because it's not
intrusive with the code. More often than not, the owners of the site
don't like the authors signing it; they claim that doing so may make the
site less efficient. By adding a txt file, you can prove your authorship
(not your property) in an external, fast, easy and accessible way.

http://humanstxt.org/