Provide a minimal JSON feed for the latest articles.
Accessible via `/articles.json`

We're using this to provide a "Latest links" section in frontend. We're
using JSON here for reasons of consistency with other MAS projects, e.g.
frontend and contento. Tests are more or less copied and tweaked from
the tests for the RSS feed.

This doesn't contain as much information as /articles.rss, because that
would require a bit more work/testing and it's not required for now.