Added AWS Region variable for AWS V4 Auth
AWS V4 Authentication was requiring the region of the bucket to be defined.  Otherwise using an AWS bucket would not work.  See: https://github.com/fog/fog/issues/3275  See Also: http://stackoverflow.com/questions/27091816/retrieve-buckets-objects-without-knowing-buckets-region-with-aws-s3-rest-api

Added the :region variable, which can be defined as an environment variable, to the code here.