Remove keys duplicated from publify_core
How does this work? I18n-tasks is configured to consider both the main
app and publify_core for locale files and key usage. Then, running
`i18n-tasks normalize` removes duplicate keys, magically only touching
those in the base app.

To do this again, it is vital that the translations in publify_core are
completely healthy!