Make gemspec work outside git
This introduces Manifest.txt files to inform the gemspecs about the
files to be packaged. This is necessary to make stop an installation
of Publify printing errors when loading the local gemspecs. Such an
installation will typically not be in a git repository, causing
`git ls-files` to complain.

Manifest files can be re-created using the manifest:create rake tasks.
The contents of these files can be checked agains the current contents
of the repository using manifest:check. The latter task is also run as
part of the default task for each gem.